Refactor repository and rewrite some parts of the code.
tor-dam is now a single binary, without the external python dependency.
When running, it will spawn a new Tor instance, and a new redis-server
instance. Their info can be retrieved with netstat(8).

The handshake logic now only checks the signature in 2/2, as the signing
in 1/2 was redundant and unnecessary.

Have fun.