Update github.com/creachadair/jrpc2 to v0.19.0.
In preparation for committing to a stable v1 release for jrpc2 package, I have
made a number of updates to the library to simplify the API a bit. These
changes are discussed in creachadair/jrpc2#46 if you would like to offer input.

This commit updates the package version used here to the latest tag as of this
writing, v0.19.0, and updates the existing use to match the changes. These
changes should not produce any functional differences.